+++ CHANGELOG.md
@@ -723,3 +723,4 @@
 * Move `Algebraic` module from `Primary` to `Paradigm`
 * Define `<:*:>` infix operator
 * Rename `Tail` constructor of `Segment` to `Rest`
+* Remove 'twosome' in favor `<:*:>`